+++ spec/support/models/book.rb
@@ -21,28 +21,24 @@ class Book < ApplicationRecord
       ActiveRecord::Associations::Preloader.new(
         records: objects,
         associations: :tags_by_callback
-      ).call.first.preloaded_records
+      ).call
+      objects.to_h { |object| [object, object.tags_by_callback] }
     }
     expose :tags_count_by_association, preload: [
       :tags_by_association,
       ->(_objects, options) { options[:expose_book_tags_count_by_association] }
-    ]
+    ] do |object, _options|
+      object.tags_by_association.size
+    end
     expose :tags_count_by_callback, preload: [
       lambda { |objects, _options|
         ActiveRecord::Associations::Preloader.new(
           records: objects,
           associations: :tags_by_callback
-        ).call.first.preloaded_records
+        ).call
+        objects.to_h { |object| [object, object.tags_by_callback.size] }
       },
       ->(_objects, options) { options[:expose_book_tags_count_by_callback] }
     ]
-
-    def tags_count_by_association
-      object.tags_by_association.size
-    end
-
-    def tags_count_by_callback
-      object.tags_by_callback.size
-    end
   end
 end